+++ src/main/java/dev/victorhleme/mobiauto/controllers/UsuarioController.java
@@ -6,6 +6,7 @@ import dev.victorhleme.mobiauto.entities.Usuario;
 import dev.victorhleme.mobiauto.filters.UsuarioFilter;
 import dev.victorhleme.mobiauto.mappers.UsuarioMapper;
 import dev.victorhleme.mobiauto.services.UsuarioService;
+import dev.victorhleme.mobiauto.services.impl.PermissionService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -24,10 +25,12 @@ public class UsuarioController {
 
     private final UsuarioService usuarioService;
     private final UsuarioMapper usuarioMapper;
+    private final PermissionService permissionService;
 
     @PostMapping
     public ResponseEntity<?> save(@RequestBody UsuarioCreationDto usuarioDto) {
         log.debug("Creating usuario");
+        permissionService.checkAuthority("PERM_USUARIO_CREATE_" + usuarioDto.getRevendaId());
         Usuario usuario = usuarioService.save(usuarioDto);
         final URI uri = getUri(usuario.getId());
         return ResponseEntity.created(uri).build();
@@ -36,12 +39,14 @@ public class UsuarioController {
     @GetMapping("/{id:[0-9]+}")
     public ResponseEntity<UsuarioDto> findById(@PathVariable final Long id) {
         log.debug("Finding usuario by id: {}", id);
+        permissionService.checkAuthority("PERM_USUARIO_READ_" + usuarioService.getRevendaIdFromUser(id));
         return ResponseEntity.ok(usuarioMapper.from(usuarioService.findById(id)));
     }
 
     @GetMapping
     public ResponseEntity<Page<UsuarioDto>> findAllPaginated(UsuarioFilter filter) {
         log.debug("Finding usuarios by filter");
+        permissionService.checkAuthority("PERM_USUARIO_READ_" + filter.getRevendaId());
         return ResponseEntity.ok(usuarioService.findAll(filter).map(usuarioMapper::from));
     }
 
@@ -51,12 +56,14 @@ public class UsuarioController {
         @RequestBody UsuarioDto dto
     ) {
         log.debug("Updating usuario with id: {}", id);
+        permissionService.checkAuthority("PERM_USUARIO_MANAGE_" + usuarioService.findById(id).getRevenda().getId());
         return ResponseEntity.ok(usuarioMapper.from(usuarioService.update(dto)));
     }
 
     @DeleteMapping("/{id:[0-9]+}")
     public ResponseEntity<?> delete(@PathVariable("id") Long id) {
         log.debug("Deleting usuario with id: {}", id);
+        permissionService.checkAuthority("PERM_USUARIO_MANAGE_" + usuarioService.findById(id).getRevenda().getId());
         usuarioService.delete(id);
         return ResponseEntity.noContent().build();
     }
